move the component to pmtc so that we dont have to create a new esm
