+++ packages/esm-ohri-pmtct-app/src/ptracker-reports/ptracker-report-app-menu-link.component.tsx
@@ -1,12 +1,12 @@
 import { ConfigurableLink } from '@openmrs/esm-framework';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import { ptrackerdashboardPath } from './constants';
+import { ptrackerdashboardPath } from '../constants';
 
 const PtrackerDashboardNavigation = () => {
   const { t } = useTranslation();
 
-  return <ConfigurableLink to={ptrackerdashboardPath}>{t('ptrackerReports', 'Ptracker Reports')}</ConfigurableLink>;
+  return <ConfigurableLink to={ptrackerdashboardPath}>{t('ptrackerReports', 'PTracker Reports')}</ConfigurableLink>;
 };
 
 export default PtrackerDashboardNavigation;
