--- packages/esm-ptracker-report-app/package.json
@@ -1,55 +1,0 @@
-{
-    "name": "@ohri/openmrs-esm-ohri-ptracker-report-app",
-    "version": "2.3.6",
-    "description": "Reporting for Namibia MNCH OpenMRS HIV Reference Implementation (OHRI)",
-    "browser": "dist/openmrs-esm-ohri-form-render-app.js",
-    "main": "src/index.ts",
-    "license": "MIT",
-    "homepage": "https://github.com/UCSF-IGHS/openmrs-esm-ohri#readme",
-    "keywords": [
-        "mnch",
-        "ptracker",
-        "ohri"
-    ],
-    "bugs": {
-        "url": "https://github.com/UCSF-IGHS/openmrs-esm-ohri/issues"
-    },
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/UCSF-IGHS/openmrs-esm-ohri.git"
-    },
-    "scripts": {
-        "start": "openmrs develop --backend https://openmrs-staging.globalhealthapp.net",
-        "serve": "webpack serve --mode development",
-        "debug": "npm run serve",
-        "build": "webpack --mode production",
-        "analyze": "webpack --mode=production --env.analyze=true",
-        "docker": "docker run -p 8080:8080 -p 8081:8081 histacohri/docker-openmrs-esm-ohri:1.0 npx openmrs develop --backend https://openmrs-staging.globalhealthapp.net",
-        "docker:dev": "docker compose up",
-        "lint": "eslint src --ext tsx",
-        "typescript": "tsc",
-        "prepublishOnly": "npm run build",
-        "extract-translations": "i18next 'src/**/*.component.tsx'"
-    },
-    "browserslist": [
-        "extends browserslist-config-openmrs"
-    ],
-    "publishConfig": {
-        "access": "public"
-    },
-    "dependencies": {
-        "@carbon/react": "^1.13.0",
-        "@openmrs/openmrs-form-engine-lib": "2.0.0-pre.1260",
-        "ace-builds": "1.34.1",
-        "react-ace": "^9.4.4"
-    },
-    "peerDependencies": {
-        "@openmrs/esm-framework": "5.x",
-        "dayjs": "1.x",
-        "react": "18.x",
-        "react-i18next": "11.x"
-    },
-    "devDependencies": {
-        "webpack": "^5.88.2"
-    }
-}
