--- packages/esm-ptracker-report-app/src/routes.json
@@ -1,23 +1,0 @@
-{
-  "$schema": "https://json.openmrs.org/routes.schema.json",
-  "backendDependencies": {
-    "fhir2": "^1.2.0",
-    "webservices.rest": "^2.24.0"
-  },
-    "pages": [
-      {
-        "component": "PtrackerReport",
-        "route": "ptracker-report"
-      }
-    ],
-    "extensions": [
-      {
-        "name": "version-two-link",
-        "slot": "app-menu-slot",
-        "component": "versionTwoNavLink",
-        "online": true,
-        "offline": true
-      }
-    ]
-  }
-
