--- packages/esm-ptracker-report-app/translations/en.json
@@ -1,5 +1,0 @@
-{
-
-    "ptrackerReports": "Ptracker Reports"
-  }
-
