--- packages/esm-ptracker-report-app/tsconfig.json
@@ -1,5 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "include": ["src/**/*"],
-    "exclude": ["src/**/*.test.tsx", "src/**/*.outdated.tsx"]
-  }
